Extract entity context menu for easier reuse
