+++ components/editor/entity-editor.tsx
@@ -24,7 +24,7 @@ import { Skeleton } from "@/components/ui/skeleton"
 import { InternalEntityHint } from "@/components/editor/hints/internal-entity-hint"
 import { ActionButton } from "@/components/actions/action-buttons"
 import { useGoToFileExplorer, useGoToGraph } from "@/lib/hooks"
-import { EntityBadge } from "../entity-badge"
+import { EntityBadge } from "../entity/entity-badge"
 import { useEntityEditorTabs } from "@/lib/state/entity-editor-tabs-state"
 import { useShallow } from "zustand/react/shallow"
 import { TypeSelectModal } from "@/components/modals/type-select-modal"
