+++ components/editor/select-reference-modal.tsx
@@ -12,7 +12,7 @@ import { Checkbox } from "@/components/ui/checkbox"
 import { CrateDataContext } from "@/components/providers/crate-data-provider"
 import { getEntityDisplayName, isRoCrateMetadataEntity, isRootEntity, toArray } from "@/lib/utils"
 import { Skeleton } from "@/components/ui/skeleton"
-import { EntityIcon } from "@/components/entity-icon"
+import { EntityIcon } from "@/components/entity/entity-icon"
 import { SlimClass } from "@/lib/schema-worker/helpers"
 import { useEditorState } from "@/lib/state/editor-state"
 import { CheckedState } from "@radix-ui/react-checkbox"
