+++ components/entity/entity-badge.tsx
