+++ components/modals/global-search.tsx
@@ -9,7 +9,7 @@ import { CurlyBraces, Folder, GitFork, Library, PackageSearch } from "lucide-rea
 import { memo, useCallback, useEffect, useState } from "react"
 import { useEditorState } from "@/lib/state/editor-state"
 import { getEntityDisplayName } from "@/lib/utils"
-import { EntityIcon } from "@/components/entity-icon"
+import { EntityIcon } from "@/components/entity/entity-icon"
 import { useGoToEntityEditor, useGoToPage } from "@/lib/hooks"
 import { useActionsStore } from "@/components/providers/actions-provider"
 import { ActionCommandItem } from "@/components/actions/action-buttons"
